[admin-api-schema] Moved validate function into utils
no issue
